backend/requirements.txt
Django
django_rest_framework
numpy
sklearn